commit
EKF/UKF is an optimal filtering toolbox for Matlab. Optimal filtering is
a frequently used term for a process, in which the state of a dynamic
system is estimated through noisy and indirect measurements. This
toolbox mainly consists of Kalman filters and smoothers, which are the
most common methods used in stochastic state-space estimation. The
purpose of the toolbox is not to provide highly optimized software
package, but instead to provide a simple framework for building
proof-of-concept implementations of optimal filters and smoothers to be
used in practical applications.